Refactor: PostDetailEditPage에 있는 이벤트 핸들러 PostDetailPage로 이동
같은 api를 불러오고 적용시키는 부분에서 PostDetail 페이지와 PostDetailEdit 페이지가 따로 존재할 필요가 없다고 생각.
PostDetailEdit 파일은 없애는 것으로 하고 필요한 함수 이동
