--- src/component/pages/PostDetailEdit.module.css
@@ -1,130 +1,0 @@
-ul li {
-  list-style: none;
-}
-
-.cardListContainer {
-  display: flex;
-  flex-direction: column;
-  align-items: flex-end;
-  gap: 11px;
-  margin: 63px auto 113px;
-  max-width: 1200px;
-}
-
-.cardList {
-  display: grid;
-  justify-content: center;
-  grid-template-columns: repeat(3, 384px);
-  gap: 28px 24px;
-}
-
-@media (max-width: 1248px) {
-  .cardListContainer {
-    padding: 0 24px;
-  }
-
-  .cardList {
-    grid-template-columns: repeat(3, 1fr);
-    width: 100%;
-  }
-}
-
-@media (max-width: 1024px) {
-  .cardList {
-    grid-template-columns: repeat(2, 1fr);
-  }
-}
-
-@media (max-width: 767px) {
-  .cardList {
-    grid-template-columns: 384px;
-  }
-}
-
-/* 임시 CSS */
-.cardContainer {
-  display: flex;
-  flex-direction: column;
-  gap: 16px;
-  padding: 28px 24px 24px 24px;
-  height: 280px;
-  border-radius: 16px;
-  background: var(--white, #fff);
-  box-shadow: 0px 2px 12px 0px rgba(0, 0, 0, 0.08);
-}
-
-.profileContainer {
-  display: flex;
-  justify-content: space-between;
-}
-
-.profile {
-  display: flex;
-  gap: 14px;
-}
-
-.img {
-  width: 56px;
-  height: 56px;
-  border-radius: 100px;
-  border: 1px solid var(--gray-200, #eee);
-}
-
-.nameAndBadgeContainer {
-  display: flex;
-  flex-direction: column;
-  gap: 6px;
-}
-
-.nameContainer {
-  display: flex;
-  gap: 6px;
-  font-size: 20px;
-  line-height: 24px;
-}
-
-.name {
-  font-weight: 700;
-}
-
-.badge {
-  padding: 0px 8px;
-  border-radius: 4px;
-  background: var(--Purple-100, #f8f0ff);
-  width: 41px;
-  color: var(--Purple-600, #9935ff);
-  font-size: var(--xs);
-  line-height: 20px;
-  letter-spacing: -0.07px;
-}
-
-.line {
-  height: 1px;
-  background: var(--gray-200, #eee);
-}
-
-.textAreaContainer {
-  flex-grow: 1;
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  padding-top: 16px;
-  box-shadow: inset 0 1px var(--gray-200);
-}
-
-.textBox {
-  height: 106px;
-  overflow: hidden;
-  color: var(--gray-600, #4a4a4a);
-  text-overflow: ellipsis;
-  font-size: var(--l);
-  line-height: 28px;
-  letter-spacing: -0.18px;
-}
-
-.date {
-  font-size: 12px;
-  line-height: 18px;
-  letter-spacing: -0.06px;
-  color: var(--gray-400, #999);
-}
